+++ epsilonsecurity/app/models/EventId.java
@@ -4,7 +4,7 @@ import javax.annotation.Nonnull;
 
 public class EventId extends DatabaseId {
 
-    public EventId(@Nonnull long rawId) {
+    public EventId(@Nonnull Integer rawId) {
         super(rawId);
     }
